Functionality/authorization+authentication refactor/co 164

+++ controllers/courses/courseController.js
@@ -10,7 +10,7 @@ const listAllCourses = async (req, res) => {
   }
 }
 
-const findByID = async (req, res) => {
+const findCourseByID = async (req, res) => {
   try {
 
     let ID = req.params.id;
@@ -20,7 +20,7 @@ const findByID = async (req, res) => {
     });
 
     if(course){
-      res.status(STATUS_CODES.STATUS_OK).json({course});
+      res.status(STATUS_CODES.STATUS_OK).json(course);
     }else
       res.status(STATUS_CODES.NOT_FOUND).json({message: 'Course not found'});
   } catch (error) {
@@ -31,14 +31,12 @@ const findByID = async (req, res) => {
 const createCourse = async (req, res) => {
 
   try {
-
     const tempCourse = {
       name: req.body.name,
       description: req.body.description,
       espb: req.body.espb,
       professorId: req.body.professorId,
     }
-    
     const course = await Course.create(tempCourse);
     res.status(STATUS_CODES.STATUS_OK).json(course);
   } catch (error) {
@@ -50,11 +48,21 @@ const deleteCourse = async (req, res) => {
 
   try {
     let ID = req.params.id;
-    const course = await Course.destroy({
-      where: {id: ID},
+    const course = await Course.findOne({
+      where: {id: ID}
     });
     if (course) {
-      res.status(STATUS_CODES.STATUS_OK).json({message: 'Course deleted!'});
+      // console.log(req.userData.id);
+      // console.log(course.professorId);
+      if(req.userData.id == course.professorId){
+        await Course.destroy({
+          where: {id: ID}
+        });
+        res.status(STATUS_CODES.STATUS_OK).json({message: 'Course deleted!'});
+      }
+      else{
+        res.sendStatus(STATUS_CODES.FORBIDDEN);
+      }
     } else {
       res.status(STATUS_CODES.NOT_FOUND).json({message: 'Course not found'});
     }
@@ -65,20 +73,29 @@ const deleteCourse = async (req, res) => {
 
 const updateCourse = async (req, res) => {
 
-  try {
 
+  try {
     let ID = req.params.id;
-    
-    const course = await Course.update({
-      name: req.body.name,
-      description: req.body.description,
-      espb: req.body.espb,
-      professorId: req.body.professorId,
-    }, {
-      where: {id: ID},
+    const course = await Course.findOne({
+      where: {id: ID}
     });
     if (course) {
-      res.status(STATUS_CODES.STATUS_OK).json({message: 'Course updated!'});
+      //console.log(req.userData.id);
+      //console.log(course.professorId);
+      if(req.userData.id == course.professorId){
+        await Course.update({
+          name: req.body.name,
+          description: req.body.description,
+          espb: req.body.espb,
+          professorId: req.body.professorId,
+        }, {
+          where: {id: ID},
+        });
+        res.status(STATUS_CODES.STATUS_OK).json({message:"Course updated!"});
+      }
+      else{
+        res.sendStatus(STATUS_CODES.FORBIDDEN);
+      }
     } else {
       res.status(STATUS_CODES.NOT_FOUND).json({message: 'Course not found'});
     }
@@ -89,8 +106,29 @@ const updateCourse = async (req, res) => {
 
 module.exports = {
   listAllCourses,
-  findByID,
+  findCourseByID,
   createCourse,
   deleteCourse,
   updateCourse
 }
+
+//   try {
+
+//     let ID = req.params.id;
+    
+//     const course = await Course.update({
+//       name: req.body.name,
+//       description: req.body.description,
+//       espb: req.body.espb,
+//       professorId: req.body.professorId,
+//     }, {
+//       where: {id: ID},
+//     });
+//     if (course) {
+//       res.status(STATUS_CODES.STATUS_OK).json({message: 'Course updated!'});
+//     } else {
+//       res.status(STATUS_CODES.NOT_FOUND).json({message: 'Course not found'});
+//     }
+//   } catch (error) {
+//     res.status(STATUS_CODES.INTERNAL_SERVER_ERROR).json(error.message);
+//   }
